move scan + data stuff to own module
